+++ trino-lb/src/http_server/mod.rs
@@ -62,6 +62,7 @@ pub async fn start_http_server(
     metrics: Arc<Metrics>,
 ) -> Result<(), Error> {
     let tls_config = config.trino_lb.tls.clone();
+    let ports_config = config.trino_lb.ports.clone();
     let app_state = Arc::new(AppState {
         config,
         persistence,
@@ -75,7 +76,7 @@ pub async fn start_http_server(
         .route("/", get(|| async { Redirect::permanent("/metrics") }))
         .route("/metrics", get(metrics::get))
         .with_state(Arc::clone(&app_state));
-    let listen_addr = SocketAddr::from((Ipv6Addr::UNSPECIFIED, 9090));
+    let listen_addr = SocketAddr::from((Ipv6Addr::UNSPECIFIED, ports_config.metrics));
     info!(%listen_addr, "Starting metrics exporter");
 
     let handle = Handle::new();
@@ -123,7 +124,7 @@ pub async fn start_http_server(
 
     if tls_config.enabled {
         // Start https server
-        let listen_addr = SocketAddr::from((Ipv6Addr::UNSPECIFIED, 8443));
+        let listen_addr = SocketAddr::from((Ipv6Addr::UNSPECIFIED, ports_config.https));
         info!(%listen_addr, "Starting server");
 
         let cert_pem_file = tls_config.cert_pem_file.context(CertsMissingSnafu)?;
@@ -142,7 +143,7 @@ pub async fn start_http_server(
             .context(StartHttpServerSnafu)?;
     } else {
         // Start http server
-        let listen_addr = SocketAddr::from((Ipv6Addr::UNSPECIFIED, 8080));
+        let listen_addr = SocketAddr::from((Ipv6Addr::UNSPECIFIED, ports_config.http));
         info!(%listen_addr, "Starting server");
 
         axum_server::bind(listen_addr)